test: move e2e health checking tests out of end2end_test.go
