+++ test/healthcheck_test.go
@@ -31,7 +31,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/connectivity"
 	"google.golang.org/grpc/credentials/insecure"
-	_ "google.golang.org/grpc/health"
+	"google.golang.org/grpc/health"
 	healthgrpc "google.golang.org/grpc/health/grpc_health_v1"
 	healthpb "google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/internal"
@@ -866,4 +866,314 @@ func (s) TestHealthCheckChannelzCountingCallFailure(t *testing.T) {
 	}); err != nil {
 		t.Fatal(err)
 	}
+}
+
+// healthCheck is a helper function to make a unary health check RPC and return
+// the response.
+func healthCheck(d time.Duration, cc *grpc.ClientConn, service string) (*healthpb.HealthCheckResponse, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), d)
+	defer cancel()
+	hc := healthgrpc.NewHealthClient(cc)
+	return hc.Check(ctx, &healthpb.HealthCheckRequest{Service: service})
+}
+
+// verifyHealthCheckStatus is a helper function to verify that the current
+// health status of the service matches the one passed in 'wantStatus'.
+func verifyHealthCheckStatus(t *testing.T, d time.Duration, cc *grpc.ClientConn, service string, wantStatus healthpb.HealthCheckResponse_ServingStatus) {
+	t.Helper()
+	resp, err := healthCheck(d, cc, service)
+	if err != nil {
+		t.Fatalf("Health/Check(_, _) = _, %v, want _, <nil>", err)
+	}
+	if resp.Status != wantStatus {
+		t.Fatalf("Got the serving status %v, want %v", resp.Status, wantStatus)
+	}
+}
+
+// verifyHealthCheckErrCode is a helper function to verify that a unary health
+// check RPC returns an error with a code set to 'wantCode'.
+func verifyHealthCheckErrCode(t *testing.T, d time.Duration, cc *grpc.ClientConn, service string, wantCode codes.Code) {
+	t.Helper()
+	if _, err := healthCheck(d, cc, service); status.Code(err) != wantCode {
+		t.Fatalf("Health/Check() got errCode %v, want %v", status.Code(err), wantCode)
+	}
+}
+
+// newHealthCheckStream is a helper function to start a health check streaming
+// RPC, and returns the stream.
+func newHealthCheckStream(t *testing.T, cc *grpc.ClientConn, service string) (healthgrpc.Health_WatchClient, context.CancelFunc) {
+	t.Helper()
+	ctx, cancel := context.WithCancel(context.Background())
+	hc := healthgrpc.NewHealthClient(cc)
+	stream, err := hc.Watch(ctx, &healthpb.HealthCheckRequest{Service: service})
+	if err != nil {
+		t.Fatalf("hc.Watch(_, %v) failed: %v", service, err)
+	}
+	return stream, cancel
+}
+
+// healthWatchChecker is a helper function to verify that the next health
+// status returned on the given stream matches the one passed in 'wantStatus'.
+func healthWatchChecker(t *testing.T, stream healthgrpc.Health_WatchClient, wantStatus healthpb.HealthCheckResponse_ServingStatus) {
+	t.Helper()
+	response, err := stream.Recv()
+	if err != nil {
+		t.Fatalf("stream.Recv() failed: %v", err)
+	}
+	if response.Status != wantStatus {
+		t.Fatalf("got servingStatus %v, want %v", response.Status, wantStatus)
+	}
+}
+
+// TestHealthCheckSuccess invokes the unary Check() RPC on the health server in
+// a successful case.
+func (s) TestHealthCheckSuccess(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthCheckSuccess(t, e)
+	}
+}
+
+func testHealthCheckSuccess(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	defer te.tearDown()
+
+	verifyHealthCheckErrCode(t, 1*time.Second, te.clientConn(), defaultHealthService, codes.OK)
+}
+
+// TestHealthCheckFailure invokes the unary Check() RPC on the health server
+// with an expired context and expects the RPC to fail.
+func (s) TestHealthCheckFailure(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthCheckFailure(t, e)
+	}
+}
+
+func testHealthCheckFailure(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.declareLogNoise(
+		"Failed to dial ",
+		"grpc: the client connection is closing; please retry",
+	)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	defer te.tearDown()
+
+	verifyHealthCheckErrCode(t, 0*time.Second, te.clientConn(), defaultHealthService, codes.DeadlineExceeded)
+	awaitNewConnLogOutput()
+}
+
+// TestHealthCheckOff makes a unary Check() RPC on the health server where the
+// health status of the defaultHealthService is not set, and therefore expects
+// an error code 'codes.NotFound'.
+func (s) TestHealthCheckOff(t *testing.T) {
+	for _, e := range listTestEnv() {
+		// TODO(bradfitz): Temporarily skip this env due to #619.
+		if e.name == "handler-tls" {
+			continue
+		}
+		testHealthCheckOff(t, e)
+	}
+}
+
+func testHealthCheckOff(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+
+	verifyHealthCheckErrCode(t, 1*time.Second, te.clientConn(), defaultHealthService, codes.NotFound)
+}
+
+// TestHealthWatchMultipleClients makes a streaming Watch() RPC on the health
+// server with multiple clients and expects the same status on both streams.
+func (s) TestHealthWatchMultipleClients(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthWatchMultipleClients(t, e)
+	}
+}
+
+func testHealthWatchMultipleClients(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+
+	cc := te.clientConn()
+	stream1, cf1 := newHealthCheckStream(t, cc, defaultHealthService)
+	defer cf1()
+	healthWatchChecker(t, stream1, healthpb.HealthCheckResponse_SERVICE_UNKNOWN)
+
+	stream2, cf2 := newHealthCheckStream(t, cc, defaultHealthService)
+	defer cf2()
+	healthWatchChecker(t, stream2, healthpb.HealthCheckResponse_SERVICE_UNKNOWN)
+
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_NOT_SERVING)
+	healthWatchChecker(t, stream1, healthpb.HealthCheckResponse_NOT_SERVING)
+	healthWatchChecker(t, stream2, healthpb.HealthCheckResponse_NOT_SERVING)
+}
+
+// TestHealthWatchSameStatusmakes a streaming Watch() RPC on the health server
+// and makes sure that the health status of the server is as expected after
+// multiple calls to SetServingStatus with the same status.
+func (s) TestHealthWatchSameStatus(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthWatchSameStatus(t, e)
+	}
+}
+
+func testHealthWatchSameStatus(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+
+	stream, cf := newHealthCheckStream(t, te.clientConn(), defaultHealthService)
+	defer cf()
+
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_SERVICE_UNKNOWN)
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_SERVING)
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_NOT_SERVING)
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_NOT_SERVING)
+}
+
+// TestHealthWatchServiceStatusSetBeforeStartingServer starts a health server
+// on which the health status for the defaultService is set before the gRPC
+// server is started, and expects the correct health status to be returned.
+func (s) TestHealthWatchServiceStatusSetBeforeStartingServer(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthWatchSetServiceStatusBeforeStartingServer(t, e)
+	}
+}
+
+func testHealthWatchSetServiceStatusBeforeStartingServer(t *testing.T, e env) {
+	hs := health.NewServer()
+	te := newTest(t, e)
+	te.healthServer = hs
+	hs.SetServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+
+	stream, cf := newHealthCheckStream(t, te.clientConn(), defaultHealthService)
+	defer cf()
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_SERVING)
+}
+
+// TestHealthWatchDefaultStatusChange verifies the simple case where the
+// service starts off with a SERVICE_UNKNOWN status (because SetServingStatus
+// hasn't been called yet) and then moves to SERVING after SetServingStatus is
+// called.
+func (s) TestHealthWatchDefaultStatusChange(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthWatchDefaultStatusChange(t, e)
+	}
+}
+
+func testHealthWatchDefaultStatusChange(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+
+	stream, cf := newHealthCheckStream(t, te.clientConn(), defaultHealthService)
+	defer cf()
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_SERVICE_UNKNOWN)
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_SERVING)
+}
+
+// TestHealthWatchSetServiceStatusBeforeClientCallsWatch verifies the case
+// where the health status is set to SERVING before the client calls Watch().
+func (s) TestHealthWatchSetServiceStatusBeforeClientCallsWatch(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthWatchSetServiceStatusBeforeClientCallsWatch(t, e)
+	}
+}
+
+func testHealthWatchSetServiceStatusBeforeClientCallsWatch(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	defer te.tearDown()
+
+	stream, cf := newHealthCheckStream(t, te.clientConn(), defaultHealthService)
+	defer cf()
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_SERVING)
+}
+
+// TestHealthWatchOverallServerHealthChange verifies setting the overall status
+// of the server by using the empty service name.
+func (s) TestHealthWatchOverallServerHealthChange(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthWatchOverallServerHealthChange(t, e)
+	}
+}
+
+func testHealthWatchOverallServerHealthChange(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+
+	stream, cf := newHealthCheckStream(t, te.clientConn(), "")
+	defer cf()
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_SERVING)
+	te.setHealthServingStatus("", healthpb.HealthCheckResponse_NOT_SERVING)
+	healthWatchChecker(t, stream, healthpb.HealthCheckResponse_NOT_SERVING)
+}
+
+// TestUnknownHandler verifies that an expected error is returned (by setting
+// the unknownHandler on the server) for a service which is not exposed to the
+// client.
+func (s) TestUnknownHandler(t *testing.T) {
+	// An example unknownHandler that returns a different code and a different
+	// method, making sure that we do not expose what methods are implemented to
+	// a client that is not authenticated.
+	unknownHandler := func(srv interface{}, stream grpc.ServerStream) error {
+		return status.Error(codes.Unauthenticated, "user unauthenticated")
+	}
+	for _, e := range listTestEnv() {
+		// TODO(bradfitz): Temporarily skip this env due to #619.
+		if e.name == "handler-tls" {
+			continue
+		}
+		testUnknownHandler(t, e, unknownHandler)
+	}
+}
+
+func testUnknownHandler(t *testing.T, e env, unknownHandler grpc.StreamHandler) {
+	te := newTest(t, e)
+	te.unknownHandler = unknownHandler
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+	verifyHealthCheckErrCode(t, 1*time.Second, te.clientConn(), "", codes.Unauthenticated)
+}
+
+// TestHealthCheckServingStatus makes a streaming Watch() RPC on the health
+// server and verifies a bunch of health status transitions.
+func (s) TestHealthCheckServingStatus(t *testing.T) {
+	for _, e := range listTestEnv() {
+		testHealthCheckServingStatus(t, e)
+	}
+}
+
+func testHealthCheckServingStatus(t *testing.T, e env) {
+	te := newTest(t, e)
+	te.enableHealthServer = true
+	te.startServer(&testServer{security: e.security})
+	defer te.tearDown()
+
+	cc := te.clientConn()
+	verifyHealthCheckStatus(t, 1*time.Second, cc, "", healthpb.HealthCheckResponse_SERVING)
+	verifyHealthCheckErrCode(t, 1*time.Second, cc, defaultHealthService, codes.NotFound)
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	verifyHealthCheckStatus(t, 1*time.Second, cc, defaultHealthService, healthpb.HealthCheckResponse_SERVING)
+	te.setHealthServingStatus(defaultHealthService, healthpb.HealthCheckResponse_NOT_SERVING)
+	verifyHealthCheckStatus(t, 1*time.Second, cc, defaultHealthService, healthpb.HealthCheckResponse_NOT_SERVING)
 }
